requirements.txt
numpy==1.26.4
pandas==2.1.4
scipy==1.10.1
scikit-learn==1.3.2
joblib==1.3.2
tqdm==4.66.4
matplotlib==3.8.2
seaborn==0.13.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
transformers==4.42.0
pypdf==4.2.0
chromadb==0.5.0
pyarrow==15.0.2
langchain==0.2.12
langchain-core>=0.2.27,<0.3.0
langchain-community==0.2.11
langchain-text-splitters==0.2.2
langchain-chroma==0.1.4
langchain-huggingface==0.0.3
neo4j==5.20.0
lightgbm==4.1.0
imbalanced-learn==0.11.0
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.35.3
google-generativeai==0.3.2